[CIDMDEV-5133] - Migration Support - Read-Only Groups
- Protected the following services by flag 'feature.migration.read.only.mode.enabled'
(returns 503 when read-only enabled) :

1. POST v2.0/RAX-GRPADM/groups
2. PUT v2.0/RAX-GRPADM/groups/{groupId}
3. DELETE v2.0/RAX-GRPADM/groups/{groupId}

Users can still be added to and removed from groups. In LDAP this updates the user object, NOT the group.

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -979,7 +979,11 @@ public class Cloud20VersionResource {
             @Context UriInfo uriInfo,
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             Group group) {
-        return cloud20Service.addGroup(httpHeaders, uriInfo, authToken, group).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addGroup(httpHeaders, uriInfo, authToken, group).build();
+        }
     }
 
     @GET
@@ -1012,7 +1016,11 @@ public class Cloud20VersionResource {
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("groupId") String groupId,
             Group group) {
-        return cloud20Service.updateGroup(httpHeaders, authToken, groupId, group).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.updateGroup(httpHeaders, authToken, groupId, group).build();
+        }
     }
 
     @DELETE
@@ -1021,7 +1029,11 @@ public class Cloud20VersionResource {
             @Context HttpHeaders httpHeaders,
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("groupId") String groupId) {
-        return cloud20Service.deleteGroup(httpHeaders, authToken, groupId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deleteGroup(httpHeaders, authToken, groupId).build();
+        }
     }
 
     @GET
